+++ lib/src/parsers/token_stream/token_stream_parser.dart
@@ -70,8 +70,7 @@ class TokenStreamParser {
         state.position = $pos;
         return null;
       } else {
-        final $val = Grammar(globals: g, members: m, productions: p);
-        return Ok($val);
+        return Ok(Grammar(globals: g, members: m, productions: p));
       }
     } else {
       state.ch = $c;
@@ -198,8 +197,7 @@ class TokenStreamParser {
             parseS(state);
             final end = state.position;
             final src = state.substring(start, end).trimRight();
-            final $val = Production(expression: e, name: i, sourceCode: src, type: t);
-            return Ok($val);
+            return Ok(Production(expression: e, name: i, sourceCode: src, type: t));
           } else {
             state.ch = $c;
             state.position = $pos;
@@ -239,8 +237,7 @@ class TokenStreamParser {
     if ($orderedChoice != null) {
       final e = $orderedChoice.$1;
       e.sourceCode = state.substring(pos, state.position).trimRight();
-      final $val = e;
-      return Ok($val);
+      return Ok(e);
     } else {
       state.errorExpected('expression');
       return null;
@@ -311,8 +308,7 @@ class TokenStreamParser {
           break;
         }
       }
-      final $val = OrderedChoiceExpression(expressions: l);
-      return Ok($val);
+      return Ok(OrderedChoiceExpression(expressions: l));
     } else {
       return null;
     }
@@ -346,8 +342,7 @@ class TokenStreamParser {
         final $errorHandler = parseErrorHandler(state);
         final h = $errorHandler?.$1;
         e.sourceCode = state.substring(pos, state.position).trimRight();
-        final Expression $val = e..errorHandler = h;
-        $list.add($val);
+        $list.add(e..errorHandler = h);
         continue;
       } else {
         break;
@@ -355,11 +350,10 @@ class TokenStreamParser {
     }
     if ($list.isNotEmpty) {
       final n = $list;
-      final $val1 = SequenceExpression(expressions: n);
-      final e = $val1;
+      final $val = SequenceExpression(expressions: n);
+      final e = $val;
       e.sourceCode = state.substring(pos, state.position).trimRight();
-      final $val2 = e;
-      return Ok($val2);
+      return Ok(e);
     } else {
       return null;
     }
@@ -428,8 +422,7 @@ class TokenStreamParser {
         if ($res1 != null) {
           final e = $res1.$1;
           e.semanticValue = n;
-          final $val = e;
-          return Ok($val);
+          return Ok(e);
         } else {
           state.ch = $c;
           state.position = $pos;
@@ -481,14 +474,12 @@ class TokenStreamParser {
       final $block = parseBlock(state);
       if ($block != null) {
         final b = $block.$1;
-        final $val = PredicateExpression(negate: true, predicate: b);
-        return Ok($val);
+        return Ok(PredicateExpression(negate: true, predicate: b));
       }
       final $suffix = parseSuffix(state);
       if ($suffix != null) {
         final s = $suffix.$1;
-        final $val1 = NotPredicateExpression(expression: s);
-        return Ok($val1);
+        return Ok(NotPredicateExpression(expression: s));
       }
       state.ch = $c;
       state.position = $pos;
@@ -500,14 +491,12 @@ class TokenStreamParser {
       final $block1 = parseBlock(state);
       if ($block1 != null) {
         final b = $block1.$1;
-        final $val2 = PredicateExpression(negate: false, predicate: b);
-        return Ok($val2);
+        return Ok(PredicateExpression(negate: false, predicate: b));
       }
       final $suffix1 = parseSuffix(state);
       if ($suffix1 != null) {
         final s = $suffix1.$1;
-        final $val3 = AndPredicateExpression(expression: s);
-        return Ok($val3);
+        return Ok(AndPredicateExpression(expression: s));
       }
       state.ch = $c;
       state.position = $pos;
@@ -546,25 +535,21 @@ class TokenStreamParser {
       if (state.ch == 42) {
         state.nextChar();
         parseS(state);
-        final $val = ZeroOrMoreExpression(expression: p);
-        return Ok($val);
+        return Ok(ZeroOrMoreExpression(expression: p));
       }
       // "+"
       if (state.ch == 43) {
         state.nextChar();
         parseS(state);
-        final $val1 = OneOrMoreExpression(expression: p);
-        return Ok($val1);
+        return Ok(OneOrMoreExpression(expression: p));
       }
       // "?"
       if (state.ch == 63) {
         state.nextChar();
         parseS(state);
-        final $val2 = OptionalExpression(expression: p);
-        return Ok($val2);
+        return Ok(OptionalExpression(expression: p));
       }
-      final $val3 = p;
-      return Ok($val3);
+      return Ok(p);
     } else {
       return null;
     }
@@ -617,8 +602,7 @@ class TokenStreamParser {
     if ($productionName != null) {
       final n = $productionName.$1;
       parseS(state);
-      final $val = ProductionExpression(name: n);
-      return Ok($val);
+      return Ok(ProductionExpression(name: n));
     } else {
       return null;
     }
@@ -635,8 +619,7 @@ class TokenStreamParser {
     final $block = parseBlock(state);
     if ($block != null) {
       final b = $block.$1;
-      final $val = ActionExpression(source: b);
-      return Ok($val);
+      return Ok(ActionExpression(source: b));
     } else {
       return null;
     }
@@ -658,8 +641,7 @@ class TokenStreamParser {
     final $block = parseBlock(state);
     if ($block != null) {
       final b = $block.$1;
-      final $val = ValueExpression(source: b, valueType: t);
-      return Ok($val);
+      return Ok(ValueExpression(source: b, valueType: t));
     } else {
       state.ch = $c;
       state.position = $pos;
@@ -698,8 +680,7 @@ class TokenStreamParser {
           final $val = GroupExpression(expression: e);
           final g = $val;
           t != null ? g.type = t : null;
-          final $val1 = g;
-          return Ok($val1);
+          return Ok(g);
         } else {
           state.errorExpected(')');
           state.ch = $c;
@@ -749,8 +730,7 @@ class TokenStreamParser {
       final $str = state.substring($pos, state.position);
       final n = $str;
       parseS(state);
-      final $val = TokenExpression(name: n);
-      return Ok($val);
+      return Ok(TokenExpression(name: n));
     } else {
       state.errorExpected('token name');
       return null;
@@ -803,8 +783,7 @@ class TokenStreamParser {
                 if (state.ch == 125) {
                   state.nextChar();
                   parseS(state);
-                  final $val = WhileExpression(expression: e, range: r).setType(t);
-                  return Ok($val);
+                  return Ok(WhileExpression(expression: e, range: r).setType(t));
                 } else {
                   state.errorExpected('}');
                   state.ch = $c;
@@ -884,8 +863,7 @@ class TokenStreamParser {
         $val = null;
       }
       final n = $val;
-      final $val1 = (m, n);
-      return Ok($val1);
+      return Ok((m, n));
     } else {
       return null;
     }
@@ -1036,8 +1014,7 @@ class TokenStreamParser {
       if (state.ch == 34) {
         state.nextChar();
         parseS(state);
-        final $val = p.join();
-        return Ok($val);
+        return Ok(p.join());
       } else {
         state.errorExpected('"');
         state.ch = $c;
@@ -1103,8 +1080,7 @@ class TokenStreamParser {
       if (state.ch == 39) {
         state.nextChar();
         parseS(state);
-        final $val = p.join();
-        return Ok($val);
+        return Ok(p.join());
       } else {
         state.errorExpected('\'');
         state.ch = $c;
@@ -1186,8 +1162,7 @@ class TokenStreamParser {
             // ~{ state.error('Unterminated Unicode escape sequence'); }
             if (state.ch == 125) {
               state.nextChar();
-              final $val = String.fromCharCode(v);
-              return Ok($val);
+              return Ok(String.fromCharCode(v));
             } else {
               state.errorExpected('}');
               state.error('Unterminated Unicode escape sequence');
@@ -1208,68 +1183,57 @@ class TokenStreamParser {
       // "a"
       if (state.ch == 97) {
         state.nextChar();
-        const $val1 = '\u0007';
-        return const Ok($val1);
+        return const Ok('\u0007');
       }
       // "b"
       if (state.ch == 98) {
         state.nextChar();
-        const $val2 = '\b';
-        return const Ok($val2);
+        return const Ok('\b');
       }
       // "e"
       if (state.ch == 101) {
         state.nextChar();
-        const $val3 = '\u001B';
-        return const Ok($val3);
+        return const Ok('\u001B');
       }
       // "f"
       if (state.ch == 102) {
         state.nextChar();
-        const $val4 = '\f';
-        return const Ok($val4);
+        return const Ok('\f');
       }
       // "n"
       if (state.ch == 110) {
         state.nextChar();
-        const $val5 = '\n';
-        return const Ok($val5);
+        return const Ok('\n');
       }
       // "r"
       if (state.ch == 114) {
         state.nextChar();
-        const $val6 = '\r';
-        return const Ok($val6);
+        return const Ok('\r');
       }
       // "t"
       if (state.ch == 116) {
         state.nextChar();
-        const $val7 = '\t';
-        return const Ok($val7);
+        return const Ok('\t');
       }
       // "v"
       if (state.ch == 118) {
         state.nextChar();
-        const $val8 = '\v';
-        return const Ok($val8);
+        return const Ok('\v');
       }
       // "\\"
       if (state.ch == 92) {
         state.nextChar();
-        const $val9 = '\\';
-        return const Ok($val9);
+        return const Ok('\\');
       }
       // "\""
       if (state.ch == 34) {
         state.nextChar();
-        const $val10 = '"';
-        return const Ok($val10);
+        return const Ok('"');
       }
       // "'"
       if (state.ch == 39) {
         state.nextChar();
-        const $val11 = '\'';
-        return const Ok($val11);
+        return const Ok('\'');
       }
       state.error('Illegal escape character');
       state.ch = $c;
@@ -1307,8 +1271,7 @@ class TokenStreamParser {
     if ($ok) {
       final $str = state.substring($pos, state.position);
       final n = $str;
-      final $val = int.parse(n);
-      return Ok($val);
+      return Ok(int.parse(n));
     } else {
       state.errorExpected('decimal number');
       return null;
@@ -1344,8 +1307,7 @@ class TokenStreamParser {
       }
       final $str = state.substring($pos, state.position);
       final n = $str;
-      final $val = int.parse(n);
-      return Ok($val);
+      return Ok(int.parse(n));
     } else {
       state.errorExpected('decimal number from 1');
       return null;
@@ -1379,8 +1341,7 @@ class TokenStreamParser {
     if ($ok) {
       final $str = state.substring($pos, state.position);
       final n = $str;
-      final $val = int.parse(n, radix: 16);
-      return Ok($val);
+      return Ok(int.parse(n, radix: 16));
     } else {
       state.errorExpected('hexadecimal number');
       return null;